Move _download into utils.js
